Reviewer's comment.
